feat: 회원 정보 조회 및 수정 / 비밀번호 변경 API 구현

+++ backend/src/main/java/com/likelion/dev_community/domain/user/service/AuthService.java
@@ -2,18 +2,19 @@ package com.likelion.dev_community.domain.user.service;
 
 import com.likelion.dev_community.common.exception.CustomException;
 import com.likelion.dev_community.common.exception.ErrorCode;
-import com.likelion.dev_community.domain.user.dto.*;
+import com.likelion.dev_community.domain.user.dto.authDto.*;
 import com.likelion.dev_community.domain.user.entity.RefreshToken;
 import com.likelion.dev_community.domain.user.entity.User;
+import com.likelion.dev_community.domain.user.entity.UserStatus;
 import com.likelion.dev_community.domain.user.repository.RefreshTokenRepository;
 import com.likelion.dev_community.domain.user.repository.UserRepository;
+import com.likelion.dev_community.security.jwt.CookieProvider;
 import com.likelion.dev_community.security.jwt.JwtProvider;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.JwtException;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseCookie;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -31,10 +32,7 @@ public class AuthService {
     private final PasswordEncoder passwordEncoder;
     private final JwtProvider jwtProvider;
     private final RefreshTokenRepository refreshTokenRepository;
-
-
-    @Value("${cookie.secure}")
-    private boolean cookieSecure;
+    private final CookieProvider cookieProvider;
 
     // 회원가입
     @Transactional
@@ -65,6 +63,9 @@ public class AuthService {
         if(!passwordEncoder.matches(request.getPassword(), user.getPassword()))
             throw new CustomException(ErrorCode.INVALID_CREDENTIALS);
 
+        if(user.getStatus() == UserStatus.WITHDRAWN)
+            throw new CustomException(ErrorCode.INVALID_CREDENTIALS);
+
         refreshTokenRepository.deleteByUserId(user.getId());
 
         String accessToken = jwtProvider.createAccessToken(user.getId(), user.getUsername(), user.getNickname(), List.of(user.getRole().name()));
@@ -72,13 +73,14 @@ public class AuthService {
 
         refreshTokenRepository.save(new RefreshToken(user.getId(),refreshToken,jwtProvider.getRefreshTokenExpirationMs()));
 
-        ResponseCookie cookie = buildCookie("refreshToken",refreshToken, Duration.ofMillis(jwtProvider.getRefreshTokenExpirationMs()));
+        ResponseCookie cookie = cookieProvider.createCookie("refreshToken",refreshToken, Duration.ofMillis(jwtProvider.getRefreshTokenExpirationMs()));
 
         httpServletResponse.addHeader(HttpHeaders.SET_COOKIE, cookie.toString());
 
         return TokenResponse.of(accessToken);
     }
 
+    // 토큰 재발급
     @Transactional
     public ReissueResponse reissue(String refreshToken){
 
@@ -110,23 +112,28 @@ public class AuthService {
         return ReissueResponse.of(newAccessToken);
     }
 
+    // 로그아웃
     @Transactional
     public void logout(Long userId, HttpServletResponse httpServletResponse){
 
-        ResponseCookie cookie = buildCookie("refreshToken", null, Duration.ZERO);
+        ResponseCookie cookie = cookieProvider.clearCookie("refreshToken");
 
         httpServletResponse.addHeader(HttpHeaders.SET_COOKIE, cookie.toString());
 
         refreshTokenRepository.deleteByUserId(userId);
     }
 
-    private ResponseCookie buildCookie(String name, String value, Duration expiration){
-        return ResponseCookie.from(name, value)
-                .httpOnly(true)
-                .secure(cookieSecure)
-                .sameSite("Strict")
-                .path("/")
-                .maxAge(expiration)
-                .build();
+    // 아이디 중복 확인
+    @Transactional(readOnly = true)
+    public void checkUsername(String username){
+        if(userRepository.existsByUsername(username))
+            throw new CustomException(ErrorCode.DUPLICATE_RESOURCE, "사용중인 아이디입니다." + username);
+    }
+
+    // 닉네임 중복 확인
+    @Transactional(readOnly = true)
+    public void checkNickname(String nickname){
+        if(userRepository.existsByNickname(nickname))
+            throw new CustomException(ErrorCode.DUPLICATE_RESOURCE, "사용중인 닉네임입니다." + nickname);
     }
 }
